pyproject.toml
[tool.poetry]
name = "rgbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.0.1"
discord = "^2.1.0"
Flask = "^2.2.3"
giphy_client = "^1.0.0"
python-dotenv = "^0.21.1"
pymongo = "^4.3.3"
discord-interactions = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
